requirements.txt
django
djangorestframework
python-dotenv
requests
psycopg2-binary
celery
redis